cull legacy-pipeline
