+++ cocos/game/game.ts
@@ -34,7 +34,6 @@ import { EventTarget, AsyncDelegate, sys, macro, VERSION, cclegacy, screen, Sett
 import { input } from '../input';
 import { deviceManager, LegacyRenderMode } from '../gfx';
 import { SplashScreen } from './splash-screen';
-import { RenderPipeline } from '../rendering';
 import { Layers, Node } from '../scene-graph';
 import { builtinResMgr } from '../asset/asset-manager/builtin-res-mgr';
 import { Director, director } from './director';
@@ -1123,24 +1122,35 @@ export class Game extends EventTarget {
     }
 
     private _setupRenderPipeline (): void | Promise<void> {
+        if (cclegacy.legacy_rendering === undefined && cclegacy.rendering === undefined) {
+            return Promise.reject();
+        }
+        // Pick builtin-pipeline
         const renderPipeline = settings.querySettings(Settings.Category.RENDERING, 'renderPipeline') as string;
-        if (!renderPipeline || renderPipeline === 'ca127c79-69d6-4afd-8183-d712d7b80e14') {
+        if (cclegacy.rendering && (!renderPipeline || renderPipeline === 'ca127c79-69d6-4afd-8183-d712d7b80e14')) {
+            log('Using builtin-pipeline');
             return this._setRenderPipeline();
         }
-        return new Promise<RenderPipeline>((resolve, reject): void => {
-            assetManager.loadAny(renderPipeline, (err, asset): void => ((err || !(asset instanceof RenderPipeline))
-                ? reject(err)
-                : resolve(asset)));
-        }).then((asset): void => {
-            this._setRenderPipeline(asset);
-        }).catch((reason): void => {
-            warn(reason);
-            warn(`Failed load render pipeline: ${renderPipeline}, engine failed to initialize, will fallback to default pipeline`);
-            this._setRenderPipeline();
-        });
+        // Pick legacy-pipeline or custom pipeline
+        if (cclegacy.legacy_rendering) {
+            log('Using legacy rendering pipeline');
+            return new Promise<any>((resolve, reject): void => {
+                assetManager.loadAny(renderPipeline, (err, asset): void => ((err)
+                    ? reject(err)
+                    : resolve(asset)));
+            }).then((asset): void => {
+                this._setRenderPipeline(asset);
+            }).catch((reason): void => {
+                warn(reason);
+                warn(`Failed load render pipeline: ${renderPipeline}, engine failed to initialize, will fallback to default pipeline`);
+                this._setRenderPipeline();
+            });
+        } else {
+            return this._setRenderPipeline();
+        }
     }
 
-    private _setRenderPipeline (rppl?: RenderPipeline): void {
+    private _setRenderPipeline (rppl?: any): void {
         if (!director.root!.setRenderPipeline(rppl)) {
             this._setRenderPipeline();
         }
