+++ cocos/rendering/legacy/index.ts
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 2020-2023 Xiamen Yaji Software Co., Ltd.
+ Copyright (c) 2021-2024 Xiamen Yaji Software Co., Ltd.
 
  https://www.cocos.com/
 
@@ -22,22 +22,18 @@
  THE SOFTWARE.
 */
 
-import './deprecated';
-import * as pipeline from './define';
+export { RenderPipeline } from '../render-pipeline';
+export { RenderFlow } from '../render-flow';
+export { RenderStage } from '../render-stage';
 
-export { pipeline };
-
-export * from './pass-phase';
-
-export { PipelineSceneData } from './pipeline-scene-data';
-
-export { InstancedBuffer } from './instanced-buffer';
-export { PipelineStateManager } from './pipeline-state-manager';
-
-export { PipelineEventProcessor, PipelineEventType } from './pipeline-event';
-export { DebugView } from './debug-view';
-
-export { ReflectionProbeFlow } from './reflection-probe/reflection-probe-flow';
-export { ReflectionProbeStage } from './reflection-probe/reflection-probe-stage';
-
-export * from './custom/settings';
+export { ForwardPipeline, createDefaultPipeline } from '../forward/forward-pipeline';
+export { ForwardFlow } from '../forward/forward-flow';
+export { ForwardStage } from '../forward/forward-stage';
+export { DeferredPipeline } from '../deferred/deferred-pipeline';
+export { MainFlow } from '../deferred/main-flow';
+export { GbufferStage } from '../deferred/gbuffer-stage';
+export { LightingStage } from '../deferred/lighting-stage';
+export { BloomStage } from '../deferred/bloom-stage';
+export { PostProcessStage } from '../deferred/postprocess-stage';
+export { ShadowFlow } from '../shadow/shadow-flow';
+export { ShadowStage } from '../shadow/shadow-stage';
