+++ editor/engine-features/render-config.json
@@ -282,6 +282,12 @@
             "label": "i18n:ENGINE.features.custom_pipeline.label",
             "description": "i18n:ENGINE.features.custom_pipeline.description",
             "enginePlugin": false
+        },
+        "legacy-pipeline": {
+            "default": true,
+            "label": "i18n:ENGINE.features.legacy_pipeline.label",
+            "description": "i18n:ENGINE.features.legacy_pipeline.description",
+            "enginePlugin": false
         }
     },
     "categories": {
